django-api/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==2.2
django-cors-headers==3.11.0
django-redis==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
packaging==21.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyparsing==3.0.7
python-dotenv==0.14.0
pytz==2021.3
redis==4.1.4
# redis-server==6.0.9
six==1.16.0
sqlparse==0.4.2
wrapt==1.14.0
